djangoWebsite/requirements.txt
django
django-ninja
mysql-connector-python